--- Board/openid_store/associations/https-www.google.com-apdYYPrtQGUIj7Vp7jMTz.Kn74A-7of.WUMAiKUFp25T9TGPN0.jWQY
@@ -1,6 +1,0 @@
-version:2
-handle:1.AMlYA9U0RE62yWlpeF6RsHSOlX80kEfnVTsjq8neFLq9B8lq__u6GExJJwtQqQ
-secret:Swp9ppgiJU9mW8d61WUxgoDulzI=
-issued:1390384996
-lifetime:46800
-assoc_type:HMAC-SHA1
