+++ Board/openid_store/associations/https-www.google.com-apdYYPrtQGUIj7Vp7jMTz.Kn74A-KVFF_I4B.MnVO53LbJDinc51fUk
@@ -1,0 +1,6 @@
+version:2
+handle:1.AMlYA9WX_AWO8E_a6gV-LT_3hTPC3Ts4PQya61XCMyYsmLXTHy11T_U4EjLq5g
+secret:8ck834P69GXfB8PthEwrXIAzxdc=
+issued:1390439577
+lifetime:46800
+assoc_type:HMAC-SHA1
